Using CamelCase for all the classes names.
The only exception (and is subjective) is class_ since that one has to match the one in builtins.py. And since all of the builtin classes are in lowercase, I decided to leave it as class_ and not Class.

Another consequence of this is that the class that used to be called "function" and now is "Function" collides with the other class called "Function" from compiler.ast.
To solve this conflict I'm importing Function from compiler.ast as FunctionNode