Changed the testprogs.py script to using the module multiprocessing.
It's not customizable like tests/run.py, but in this one there is no conflict between the tests. And unlike tests/run.py, the output is shown, but only after the test has finished. This means that the output is not shown in "real time", but rather as whole after the test is completed