remove trivial dependency/jinja2 experiment
not sure if templates are a good fit for shedskin.

perhaps if another language than C++ would be supported,
and someone spends a year on untangling things..

both are unlikely.

I would have actually liked to see a Java backend in the
past (and get lots of optimizations for free).