Reverted the package name adjustments.
Since these names are used more widely, it is not helpful to optimise
them in the CMake files, and CMake should be able to understand which
packages are involved, anyway. Of course, if Conan has not done the
right thing in generating the required CMake resources, then the names
probably will need changing in numerous places.

Signed-off-by: Paul Boddie <paul@boddie.org.uk>