test_builtins_sorted: uncomment and fix
problem was 3-tuples with different types, whereas only 2-tuples
with different types are supported at the moment.

so we make the 3-tuples into 2-tuples: (a,b,c) -> ((a,b), c)

of course this is not very pretty and in practice a class with
three attributes is probably nicer, but it should work.