examples/genetic2: fix random seeding
make sure we use a wide range of seeds, so the measurements
are more stable/accurate.

a single seed doesn't work since shedskin is not compatible
with the cpython mersenne twister engine - in fact it now
uses xoshiro..

we could of course be compatible (again) in the future,
by default or using a command-line option.